Don't put unrelated code to CommandRunner
